chore: exported form validation function
