+++ src/utils/FormValidation.js
@@ -1,20 +1,25 @@
 
 
-export function validateEmail(email) 
-{
+
+
+export const validateForm = (values) => {
+    const errors = {}
     let re = /\S+@\S+\.\S+/;
-    return re.test(email);
+    if(!values.username){
+        errors.username = 'Username is required!'
+    }
+    if(!values.email){
+        errors.email = 'Email is required!'
+    } else if(!re.test(values.email)){
+        errors.email = 'Please enter a valid email!'
+    }
+    if(!values.passwordOne){
+        errors.passwordOne = 'Password is required!'
+    } else if(values.passwordOne.length < 8){
+        errors.passwordOne = 'Minimum 8 characters required!'
+    }
+    if(values.passwordTwo !== values.passwordOne){
+        errors.passwordTwo = 'Confirm password!'
+    }
+    return errors;
 }
-
-// export function passwordMatch(passwordOne, passwordTwo){
-//     let successClass = 'success';
-//     let errorClass = 'error';
-
-
-//     if(passwordOne === passwordTwo){
-//         return successClass
-//     } else {
-//         return errorClass
-//     }
-
-// }
